#1 updated Ui minor changes #2 added dynamic text #3 updated backend deals controller

+++ Backend/src/main/java/com/ecom/service/ProductService.java
@@ -25,7 +25,7 @@ public class ProductService {
 	}
 
 	public boolean deleteProductById(int productId) {
-		productRepo.deleteById(productId);
+		productRepo.deleteById((long) productId);
 		return true;
 	}
